Introduce QWindow::setOpaque()
There are windows that change their opaqueness at runtime. For example,
a panel may look opaque if a window is maximized, in other cases
translucent. This opens some optimization opportunities to the compositing
manager, for example it can skip painting the desktop background, which
can contain some dynamic contents, for example an analog clock widget,
and so on.

The only way to determine if a window is opaque or not is to check the
alpha buffer size in the requested format. This doesn't work if a window
wants to change its opaqueness as the surface format cannot change at
runtime without recreating the window.

This change introduces a new hint in QWindow that such applications can
use to tell the compositing manager that the window is opaque even though
it has an alpha channel.

Change-Id: I5c33e305aaff51c72cf98756d0a784d1a5fdd1c4