Windows: Implement Qt::ExpandedClientAreaHint/NoTitleBarBackgroundHint
Windows with the Qt::ExpandedClientAreaHint set extends the client area
to include the native titlebar area, while keeping the functionality of
systemmenu, resizing and aero snap. Also this flag indicates, that Qt
is drawing a title bar as an overlay over the extended client area.
Additionally setting Qt::NoTitleBarBackgroundHint will make the
titlebar background transparent, so that the widgets below are visible
and interactible.

Task-number: QTBUG-127634
Change-Id: Ib267d199f42a4ee90e7a51fe60a41295b82031d6
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>