Add  QtOrderingPrivate::lexicographicalCompareThreeWay
This function should behave similarly to
std::lexicographical_compare_three_way, but be available in C++17.
It is required at least to properly implement the
compareThreeWay() helper function for Qt container types.

The function requires that the contained types of the compared
ranges provide a compareThreeWay() helper function.

Similarly to std implementation, this patch also adds an overload
that takes a custom comparator object.

For now the functions are added in a private namespace, because
they are only required to implement relational operators on Qt
containers. We might want to expose them as public API later, if
needed.

Task-number: QTBUG-127095
Task-number: QTBUG-120305
Change-Id: I5b29129905b2e801ae7e470c96a7ef71e7b210d6
Reviewed-by: Marc Mutz <marc.mutz@qt.io>