Use and recognize en-POSIX as BCP47 name of the C locale
While CLDR wants to call it en-US-u-va-posix, my reading of [BCP47]
and the CLDR's relevant filename, en_US_POSIX.xml, leads me to
conclude that (once we omit the territory code as implied by
likely-subtag rules) we should represent and recognize it as en-POSIX.

[BCP47] https://www.rfc-editor.org/rfc/bcp/bcp47.txt

[ChangeLog][QtCore][QLocale] Changed the QLocale::bcp47Name() of the C
locale to "en-POSIX" to distinguish it from plain "en" (from which it
differs materially). This name is now also recognized as a name for
the C locale.

Task-number: QTBUG-81051
Change-Id: I5d2758aede88dd624733daa216a1ad8c0050112a
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>