Coin: VxWorks intel related instructions
Startup the correct emulator for each arch.

Add a health check for emulator during test start. If emulator
does not respond, restart it. Occurs in declarative tests where
tests hang and crash. After the crashes the pipes also need to be
reset, only emulator reset is not sufficient.

Read variables from WIND_CC_SYSROOT which is the VSB path, so
correct VSB is used.

Change-Id: If918ee45d4fa3a452c484a48fe9507f12f788e86
Reviewed-by: Simo Fält <simo.falt@qt.io>