moc: remove always-known parameters to (Revisioned)ConstructorData
It's a meta method so it must have a name and a return type, even though
constructors have neither. The method name of a constructor QMetaMethod
is the name of the class and the name is already required to be stored
at index 0. We just have to be careful when the class is not directly in
the global namespace, because then the first string contains the full
scope.

Prior to this change, it stored an empty string as the method return
type, but QMetaMethod::returnMetaType() already protects against reading
it, with:

    if (!mobj || methodType() == QMetaMethod::Constructor)
        return QMetaType{};

which was added in commit fa987d44417528856d5e80ed7b48ba99e19fa307
("MetaObject: Store the QMetaType of the methods") because there is no
metatype stored for this non-existent return type. This commit repeats
some of that to QMetaMethod::typeName().

Change-Id: Ia903287c2c07b1d588a9fffdaf22fff0b0d6ce9f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>