qpa: Move inLiveResize to QPlatformWindow
9122d826d25b9afab230771dc86e1ff6158e6cbc introduced support in the Cocoa
backend to provide a hint to applications to not trigger new frames on
the render thread without the main thread being the driver which can
cause painting to be out of sync with the expected size and appear
glitchy.

This problem exists on other platforms, moving the current mechanism to
QPlatformWindow allows other backends to opt-into it.

Change-Id: I03ea1354c3768f2b5feb82596ce5f0d0138c5976
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>