QList: rework comparison operators
Now that we have lexicographicalCompareThreeWay(), we can implement
the compareThreeWay() helper method in terms of this function. This
method is only available if the contained type provides its own
compareThreeWay() helper method.

However, we decided to not change the existing relational operators
for C++17 mode, and only provide operator<=>() if the contained type
also implements operator<=>() or operator<(). The reason for that is
that we do not want to introduce any potential regressions in container
comparison code.

This required extending QTypeTraits with the traits to detect support
for operator<=>(). We use the std::three_way_comparable concept for
that, because the check is only needed for C++20 code.

The implementation uses std::lexicographical_compare_three_way with
a custom QtOrderingPrivate::synthThreeWay() comparator, which is
modelled after the exposition-only synth-three-way function, as defined
by the standard [0].

[0]: https://wg21.link/expos.only.entity#2

The tests show that in C++17 mode the relational operators do not
handle the partially_ordered::unordered case correctly. That is because
operator<=() and operator>=() are implemented in terms of operator<()
and operator==(). The same problem exists for std containers.

Added the unit-tests to tst_containerapisymmetry.cpp to later reuse
them for QVarLengthArray.

[ChangeLog][QtCore][QList] QList now implements operator<=>() in
C++20 mode.

Task-number: QTBUG-120305
Change-Id: I7f6fd9c1b060449a265447ec0922088dcf521cd2
Reviewed-by: Marc Mutz <marc.mutz@qt.io>