Introduce VxKeyboard input plugin
Introduce VxKeyboard plugin which is responsible for handling keyboard
on VxWorks OS. Due to the fact that VxWorks is using its own, not fully
compatible with Unix/Linux implementation of libevdev, it requires it's
own plugin. This way, changes in evdev plugins will not be
blocked/complicated by VxWorks stalled implementation.

Implement VxKeyboard plugin and its backend, adding support for it in
eglfs.

This is a preparatory change, as there is no keyboard detection for
VxWorks yet.

Task-number: QTBUG-115777
Change-Id: I4ca5729c6f210fb7477bc31371a137071e12fdad
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>