QTest: fix debug-printing of arrays
Namely, string literals (char[n], char16_t[n]). Amends
commit 0756cc1eae5fd8981983319fef1d084762a67b8d.

I had to use the std::decay_t in that commit because the cast would fail
to compile. But I didn't check that it was correct: std::addressof()
will return the address of the first element for an array, which meant
the cast would make a pointer of the string's contents and crash.

Pick-to: 6.8
Change-Id: I701ae76be9b3340db49afffd9132b6c6dbae660b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>