moc: make the result of QtMocHelpers::metaObjectData not have UniqueType
That way, there are fewer types that get stored as global variables,
possibly reducing the number classes detailed in the debug symbol table.

            Core    Gui   Widgets
Before        84    122      208
After         45     72      110

This breaks NRVO by returning a different type than the local variable,
but since this is all running at constexpr time, it doesn't really
matter.

We need to pass the unique type to the metatype resolver, so the copyTo
functions are modified to take it as a template parameter.

Change-Id: I4083f1d1c2c44ed6dd22fffd9840da7741b723d8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>