Improve handling of localized offset names for timezones
QCalendarBackend::dateTimeToString(), in qlocale.cpp, was (either via
QDateTime or directly from QTimeZone) getting a QTimeZone::OffsetName
representation, that typically includes a UTC or GMT prefix on the
offset, then pruning that prefix and potentially kludging the rest.
Depending on locale, there may be a completely different prefix
meaning the same thing, that we should prune, but don't know enough
(in QCalBE, even in qlocale.cpp) to kludge away.

When we are using QTimeZoneLocale's data to get the OffsetName, we
actually have the data to construct the properly-formatted offset
directly and, indeed, do so as part of constructing the OffsetName.
Furthermore, QTZL's QTZP::localName() falls back, when all else fails,
on a kludged form of its OffsetName. So refactor how the latter is
obtained via a new QTZL function, zoneOffsetFormat(), with a slightly
krunky signature (to accommodate what its two callers have available
and its two implementations need), and have qlocale.cpp package its
kludges to OffsetName within a fall-back implementation of that (for
use when QTZL lacks data), while QTZP::localeName() calls a QTZL
implementation that can handle all three cases gracefully (or, at
least, without ugly kludgery).

Task-number: QTBUG-129696
Change-Id: I41da36768c56ee5d459e646893a8934f00ae3f17
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>