Http2: Delay sending client preface
We need to enable users to abort the connection to a website if, after
performing the TLS handshake, they decide not to continue. Without
sending any data.

The problem with QHttp2Connection was that it would always send
the preface as soon as the object was created, along with the
SETTINGS frame, which may in turn be seen as leaking info from
the user (disregarding that ALPN negotiations already succeeded).

This change requires us to create the first stream before the SETTINGS
gets sent, so this also rearranges all of the QHttp2Connection tests
to create a stream before waiting for the SETTINGS ACK.

Not cherry-picking because this was never considered a problem for
gRPC, but was something we worked to make happen for QNAM.

Change-Id: I84063adf3401a143046514499c2443506c60192c
Reviewed-by: Mate Barany <mate.barany@qt.io>