Android: Add OnDataChangedListerner callback interface
Adds a new interface to QtAbstractItemModel Java class that
can be used as a callback to listen to the changes of an
editable model. The model can be an implementation of
QtAbstractItemModel with an underlaying C++
QAndroidAbstractModelProxy instance, or it can be an instance
of QtAbstractItemModel that proxies and refers to
a QAbstractItemModel instance.

This also adds setOnDataChangedListener() to QtAbstractItemModel
so that it can call the interface callback function by connecting
dataChanged() signal of the native instance to a lambda that
invokes the handleDataChanged() private method of the JVM instance.

Task-number: QTBUG-129387
Task-number: QTBUG-130251
Task-number: QTBUG-130252
Change-Id: I23da9e980b5402f4dcf7446560155271ccd8809e
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>