Blacklist QUdpSocket tests which fail on VxWorks
A bunch of QUdpSocket tests fail on VxWorks due reasons which are during
investigation - many failures are related to failed `bind()` call, which
results in error suggesting some configuration problems.

We need to block building COIN for VxWorks, but some issues which result
in failures are still to be addressed. Due to this, blacklist
`tst_QUdpSocket` tests on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130078
Change-Id: Ie4e6422cc6bd787964059784a1e6f30e31f2f3dd
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>